refactor(website): name homepage sections by purpose
